frontend: Unify general and cluster settings
Accessing settings in Headlamp is confusing. This change adds the cluster
settings to the sidebar along with the general and plugin settings. When
we open a cluster, clicking the Settings (gear) icon at the top will now
take us to the cluster settings page in the home context, with a cluster
selector appearing at the top of the page.

Fixes: #1927 and #2037

Signed-off-by: Evangelos Skopelitis <eskopelitis@microsoft.com>